extract holoview result methods into separate classes
