+++ bencher/__init__.py
@@ -69,7 +69,7 @@ from .variables.parametrised_sweep import ParametrizedSweep
 from .caching import CachedParams
 from .results.bench_result import BenchResult
 from .results.panel_result import PanelResult
-from .results.holoview_result import ReduceType, HoloviewResult
+from .results.holoview_results.holoview_result import ReduceType, HoloviewResult
 from .bench_report import BenchReport, GithubPagesCfg
 from .job import Executors
 from .video_writer import VideoWriter, add_image
